+++ R/zzz_exports.R
@@ -10,10 +10,6 @@ magrittr::`%<>%`
 #' @export
 magrittr::`%$%`
 
-#' @importFrom haven as_factor
-#' @export
-haven::as_factor
-
 #' @importFrom sjlabelled get_label
 #' @export
 sjlabelled::get_label
@@ -29,7 +25,3 @@ sjlabelled::set_label
 #' @importFrom sjlabelled set_labels
 #' @export
 sjlabelled::set_labels
-
-#' @importFrom sjmisc word_wrap
-#' @export
-sjmisc::word_wrap