+++ R/zzz_exports.R
@@ -9,19 +9,3 @@ magrittr::`%<>%`
 #' @importFrom magrittr %$%
 #' @export
 magrittr::`%$%`
-
-#' @importFrom sjlabelled get_label
-#' @export
-sjlabelled::get_label
-
-#' @importFrom sjlabelled get_labels
-#' @export
-sjlabelled::get_labels
-
-#' @importFrom sjlabelled set_label
-#' @export
-sjlabelled::set_label
-
-#' @importFrom sjlabelled set_labels
-#' @export
-sjlabelled::set_labels