+++ NEWS.md
@@ -11,8 +11,8 @@
     - `nom_lambda`: Use `DescTools::Lambda`
     - `ord_gamma`: Use `DescTools::GoodmanKruskalGamma`
     - `ord_somers_d`: Use `DescTools::SomersDelta`
-    
-The `DescTools` functions are very similar to what the wrappers do, so the wrappers might be removed in the future.
+    - The `DescTools` functions are very similar to what the wrappers do, so the wrappers might be removed in the future.
+- Eliminate `lazyeval` dependency in `tadaa_int` by being better at `ggplot2`.
 
 # tadaatoolbox 0.15.0