+++ NEWS.md
@@ -12,8 +12,12 @@
     - `ord_gamma`: Use `DescTools::GoodmanKruskalGamma`
     - `ord_somers_d`: Use `DescTools::SomersDelta`
     - The `DescTools` functions are very similar to what the wrappers do, so the wrappers might be removed in the future.
+- Eliminate `vcd` dependency, also in favor of `DescTools`.
 - Eliminate `lazyeval` dependency in `tadaa_int` by being better at `ggplot2`.
 - Eliminate `dplyr` dependency by being better at R.
+- Eliminate `haven` dependency by not re-exporting `as_factor` anymore ¯\\\_(ツ)_/¯
+- Eliminare `sjmisc` depencency because why did we depend on that again?
+
 # tadaatoolbox 0.15.0
 
 - Fix error in SEM calculation in `tadaa_one_sample` for t-tests